Add a yellow star by games where it's your turn
Fixes #117
