+++ app/javascript/src/App.vue
@@ -7,8 +7,12 @@
 
 <script>
 import { apiClient } from "./router/index.js";
+import Imperial from "../lib/imperial.js";
+import { Nation } from "../lib/constants.js";
 
 import Header from "./components/Header.vue";
+
+import getGameLog from "./getGameLog.js";
 
 export default {
   name: "App",
@@ -30,12 +34,15 @@ export default {
     });
     apiClient.onUpdateGames(({ games }) => {
       this.games = games.map(game => {
+        const gameLog = getGameLog(game.log);
+        const imperialGame = Imperial.fromLog(gameLog);
         return {
           host: game.host,
           log: game.log,
           players: game.players,
           name: game.name,
-          id: game.id
+          id: game.id,
+          currentPlayer: imperialGame.currentPlayerName
         };
       });
     });
