+++ app/javascript/src/components/Star.vue
@@ -1,0 +1,16 @@
+<template>
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    height="20"
+    viewBox="0 0 51 48"
+    class="inline-block align-baseline"
+  >
+  <path fill="#ffd700" stroke="#000" d="m25,1 6,17h18l-14,11 5,17-15-10-15,10 5-17-14-11h18z"/>
+  </svg>
+</template>
+
+<script>
+export default {
+  name: "Star"
+};
+</script>
